- OutboxJobHandler, logger type fixed - refactor fetch envelopes from DB - rename OutboxHandler.cs => MainLoopOutboxHandler - refactoring OutboxHandler.ProcessAsync(), decrease code complexity - fix tests
